requirements.txt
# Abgeleitet aus den Python-Imports im Projekt.
# Installation: pip install -r requirements.txt

beautifulsoup4>=4.12
google-cloud-storage>=2.0
lxml>=4.9
psutil>=5.9
psycopg[binary]>=3.1,<4
pydantic>=2.0,<3
requests>=2.31
